Encuesta set status

+++ src/app/components/finish/finish.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { AliceService } from 'src/app/services/alice.service';
+import { LocalService } from 'src/app/services/local.service';
 
 @Component({
   selector: 'app-finish',
@@ -10,19 +12,30 @@ export class FinishComponent implements OnInit {
   loader: boolean = true;
   finish: boolean = false;
 
-  constructor() {
-    
+  token: string = '';
+
+  constructor(private alice: AliceService,
+              private localService: LocalService) {
+
+      this.localService.getToken().then(res => {
+        this.token = res;
+        this.setStatusSurvey(this.token);
+      });
    }
 
   ngOnInit(): void {
-    localStorage.clear();
-    setTimeout(function() {
-      location.href = "http://www.sitwifi.com/";
-  }, 7000);
+    
   }
 
-  hideLoading(){
-   
+  setStatusSurvey(token: string){
+    this.alice.setStatusSurvey({token_uuid: token}).subscribe(res => {   
+      if(res.code == 200){
+         localStorage.clear();
+        setTimeout(function() {
+          location.href = "http://www.sitwifi.com/";
+        }, 7000);
+      }
+    })
   }
 
 }
